【RE】[AtCoder Beginner Contest 179]F - Simplified Reversi

+++ abc179/abc179_f.cpp
@@ -38,7 +38,7 @@ bool contain(set<int> &s, int a) { return s.find(a) != s.end(); }
 typedef priority_queue<ll, vector<ll>, greater<ll> > PQ_ASK;
 
 struct Query {
-    int orient, x;
+    ll orient, x;
 };
 
 template<class T, class F>
